requirements-minimal.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Atlassian SDKs (replaces custom HTTP code)
atlassian-python-api==3.41.0

# AI/LLM Integration
anthropic==0.18.1
openai==1.12.0

# MCP Integration
mcp==0.9.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1

# Logging
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx-mock==0.1.3

